Added jsconfig.json to enable auto import update. Refactored imports for event enums
