+++ jsconfig.json
@@ -1,0 +1,7 @@
+{
+    "compilerOptions": {
+      "module": "commonjs",
+      "target": "es6"
+    },
+    "exclude": ["node_modules"]
+  }
